v0.4.8

+++ sol_kernel/fwd.py
@@ -388,7 +388,7 @@ def _forward_ptr(
     q_start = q_block * BLOCK_SIZE
     q_rows = q_start + token_offsets
     q = tl.load(
-        q_ptr + batch * sq_b + q_rows[:, None] * sq_t + head * sq_h + d_offsets[None, :],
+        q_ptr + batch * sq_b + q_rows[:, None].to(tl.int64) * sq_t + head * sq_h + d_offsets[None, :],
         mask=(q_rows < T)[:, None],
         other=0.0,
     )
@@ -458,7 +458,7 @@ def _forward_ptr(
             kv_start = block * BLOCK_SIZE
             k_rows = kv_start + token_offsets
             k = tl.load(
-                k_ptr + batch * sk_b + k_rows[:, None] * sk_t + head * sk_h + d_offsets[None, :],
+                k_ptr + batch * sk_b + k_rows[:, None].to(tl.int64) * sk_t + head * sk_h + d_offsets[None, :],
                 mask=(k_rows < T)[:, None],
                 other=0.0,
             )
@@ -473,7 +473,7 @@ def _forward_ptr(
             exact_probability = tl.math.exp2(exact_scores - new_max[:, None])
             row_sum = row_sum * alpha + tl.sum(exact_probability, axis=1)
             v = tl.load(
-                v_ptr + batch * sv_b + k_rows[:, None] * sv_t + head * sv_h + bv_offsets[None, :],
+                v_ptr + batch * sv_b + k_rows[:, None].to(tl.int64) * sv_t + head * sv_h + bv_offsets[None, :],
                 mask=(k_rows < T)[:, None],
                 other=0.0,
             )
@@ -530,7 +530,7 @@ def _forward_int8_ptr(
     q_rows = q_start + token_offsets
     q_valid = q_rows < T
     q = tl.load(
-        q_ptr + batch * sq_b + q_rows[:, None] * sq_t + head * sq_h + d_offsets[None, :],
+        q_ptr + batch * sq_b + q_rows[:, None].to(tl.int64) * sq_t + head * sq_h + d_offsets[None, :],
         mask=q_valid[:, None],
         other=0.0,
     )
@@ -624,7 +624,7 @@ def _forward_int8_ptr(
             exact_probability = tl.math.exp2(exact_scores - new_max[:, None])
             row_sum = row_sum * alpha + tl.sum(exact_probability, axis=1)
             v = tl.load(
-                v_ptr + batch * sv_b + k_rows[:, None] * sv_t + head * sv_h + bv_offsets[None, :],
+                v_ptr + batch * sv_b + k_rows[:, None].to(tl.int64) * sv_t + head * sv_h + bv_offsets[None, :],
                 mask=k_valid[:, None],
                 other=0.0,
             )
